add tasklist tsx and extension for highlighting
